+++ bin/aqueduct.dart
@@ -1,8 +1,9 @@
-import 'dart:async';
-import 'package:args/args.dart';
-import 'package:aqueduct/aqueduct.dart';
+import 'dart:io';
 
-Future<int> main(List<String> args) async {
+import 'package:aqueduct/aqueduct.dart';
+import 'package:args/args.dart';
+
+main(List<String> args) async {
   var templateCreator = new CLITemplateCreator();
   var migrationRunner = new CLIMigrationRunner();
   var setupCommand = new CLISetup();
@@ -11,21 +12,28 @@ Future<int> main(List<String> args) async {
     ..addCommand("create", templateCreator.options)
     ..addCommand("db", migrationRunner.options)
     ..addCommand("setup", setupCommand.options)
-    ..addFlag("help", abbr: "h", negatable: false, help: "Shows this documentation");
+    ..addFlag("help",
+        abbr: "h", negatable: false, help: "Shows this documentation");
 
   var values = totalParser.parse(args);
 
   if (values.command == null) {
-    print("Invalid command, options are: ${totalParser.commands.keys.join(", ")}");
-    return -1;
+    print(
+        "Invalid command, options are: ${totalParser.commands.keys.join(", ")}");
+    exitCode = 1;
+    return;
   } else if (values.command.name == "create") {
-    return await templateCreator.process(values.command);
+    exitCode = await templateCreator.process(values.command);
+    return;
   } else if (values.command.name == "db") {
-    return await migrationRunner.process(values.command);
+    exitCode = await migrationRunner.process(values.command);
+    return;
   } else if (values.command.name == "setup") {
-    return await setupCommand.process(values.command);
+    exitCode = await setupCommand.process(values.command);
+    return;
   }
 
-  print("Invalid command, options are: ${totalParser.commands.keys.join(", ")}");
-  return -1;
+  print(
+      "Invalid command, options are: ${totalParser.commands.keys.join(", ")}");
+  exitCode = 1;
 }