+++ example/templates/default/test/register_test.dart
@@ -1,6 +1,4 @@
-import 'package:test/test.dart';
-import 'package:wildfire/wildfire.dart';
-import 'mock/startup.dart';
+import 'harness/app.dart';
 
 main() {
   group("Success cases", () {
@@ -16,11 +14,16 @@ main() {
 
     test("Can create user", () async {
       var response = await (app.client.clientAuthenticatedRequest("/register")
-        ..json = {"email": "bob@stablekernel.com", "password": "foobaraxegrind12%"}).post();
+            ..json = {
+              "email": "bob@stablekernel.com",
+              "password": "foobaraxegrind12%"
+            })
+          .post();
 
-      expect(response, hasResponse(200, partial({
-        "access_token": hasLength(greaterThan(0))
-      })));
+      expect(
+          response,
+          hasResponse(
+              200, partial({"access_token": hasLength(greaterThan(0))})));
     });
   });
 
@@ -36,24 +39,36 @@ main() {
     });
     test("Trying to create existing user fails", () async {
       await (app.client.clientAuthenticatedRequest("/register")
-        ..json = {"email": "bob@stablekernel.com", "password": "someotherpassword"}).post();
+            ..json = {
+              "email": "bob@stablekernel.com",
+              "password": "someotherpassword"
+            })
+          .post();
 
       var response = await (app.client.clientAuthenticatedRequest("/register")
-        ..json = {"email": "bob@stablekernel.com", "password": "foobaraxegrind12%"}).post();
+            ..json = {
+              "email": "bob@stablekernel.com",
+              "password": "foobaraxegrind12%"
+            })
+          .post();
 
       expect(response, hasStatus(409));
     });
 
     test("Omit password fails", () async {
       var response = await (app.client.clientAuthenticatedRequest("/register")
-        ..json = {"email": "bobby.bones@stablekernel.com",}).post();
+            ..json = {
+              "email": "bobby.bones@stablekernel.com",
+            })
+          .post();
 
       expect(response, hasStatus(400));
     });
 
     test("Omit username fails", () async {
       var response = await (app.client.clientAuthenticatedRequest("/register")
-        ..json = {"password": "foobaraxegrind12%"}).post();
+            ..json = {"password": "foobaraxegrind12%"})
+          .post();
 
       expect(response, hasStatus(400));
     });